ITB-87: Moduralized service layer into crud package
